Consolidate redundant helper functions
