+++ dynestyx/utils.py
@@ -2,12 +2,13 @@ import math
 
 import diffrax as dfx
 import equinox as eqx
+import jax
 import jax.numpy as jnp
 from cd_dynamax import ContDiscreteNonlinearGaussianSSM as CDNLGSSM
 from cd_dynamax import ContDiscreteNonlinearSSM as CDNLSSM
 from jax import Array, lax
 
-from dynestyx.models import DynamicalModel
+from dynestyx.models import ContinuousTimeStateEvolution, DynamicalModel
 
 
 def flatten_draws(arr: Array) -> Array:
@@ -36,6 +37,58 @@ def flatten_draws(arr: Array) -> Array:
 type SSMType = CDNLGSSM | CDNLSSM
 
 _CONTROL_EXTEND_EPSILON = 1e-5
+
+
+def _array_has_plate_dims(
+    arr: Array | None,
+    plate_shapes: tuple[int, ...],
+    *,
+    min_suffix_ndim: int = 0,
+) -> bool:
+    """Return True if arr has leading dims exactly matching plate_shapes."""
+    if arr is None:
+        return False
+    n_plates = len(plate_shapes)
+    if arr.ndim < n_plates:
+        return False
+    for i, size in enumerate(plate_shapes):
+        if arr.shape[i] != size:
+            return False
+    return (arr.ndim - n_plates) >= min_suffix_ndim
+
+
+def _ensure_continuous_bm_dim(dynamics: DynamicalModel) -> DynamicalModel:
+    """Infer and set bm_dim when continuous dynamics were constructed in plates."""
+    if not dynamics.continuous_time:
+        return dynamics
+
+    state_evolution = dynamics.state_evolution
+    if (
+        not isinstance(state_evolution, ContinuousTimeStateEvolution)
+        or state_evolution.diffusion_coefficient is None
+        or state_evolution.bm_dim is not None
+    ):
+        return dynamics
+
+    x0 = jnp.zeros((dynamics.state_dim,))
+    u0 = None if dynamics.control_dim == 0 else jnp.zeros((dynamics.control_dim,))
+    t0 = jnp.array(0.0) if dynamics.t0 is None else jnp.asarray(dynamics.t0)
+    diffusion_shape = jax.eval_shape(
+        lambda: state_evolution.diffusion_coefficient(x0, u0, t0)
+    ).shape
+    if len(diffusion_shape) != 2:
+        raise ValueError(
+            "diffusion_coefficient must return shape (state_dim, bm_dim). "
+            f"Got shape {diffusion_shape}."
+        )
+    if int(diffusion_shape[0]) != int(dynamics.state_dim):
+        raise ValueError(
+            "diffusion_coefficient first dimension must match state_dim. "
+            f"Got diffusion_shape={diffusion_shape}, state_dim={dynamics.state_dim}."
+        )
+    inferred_bm_dim = int(diffusion_shape[1])
+    object.__setattr__(state_evolution, "bm_dim", inferred_bm_dim)
+    return dynamics
 
 
 def _should_record_field(
